requirements.txt
streamlit
langchain-google-genai
python-dotenv
requests
opencv-python
pyzbar
Pillow
